# Add caching of event_groups and pre-partition the event types