--- a/app/models/event_stream.rb
+++ b/app/models/event_stream.rb
@@ -36,8 +36,9 @@ class EventStream < ApplicationRecord
 
   after_commit :emit_notifications, :on => :create
 
-  DEFAULT_GROUP_NAME  = :other
-  DEFAULT_GROUP_LEVEL = :detail
+  DEFAULT_GROUP_NAME     = :other
+  DEFAULT_GROUP_NAME_STR = N_("Other")
+  DEFAULT_GROUP_LEVEL    = :detail
 
   def self.description
     raise NotImplementedError, "Description must be implemented in a subclass"
@@ -59,57 +60,75 @@ class EventStream < ApplicationRecord
 
   # TODO: Consider moving since this is EmsEvent specific. group, group_level and group_name exposed as a virtual columns for reports/api.
   def self.event_groups
-    core_event_groups = ::Settings.event_handling.event_groups.to_hash
-    Settings.ems.each_with_object(core_event_groups) do |(_provider_type, provider_settings), event_groups|
-      provider_event_groups = provider_settings.fetch_path(:event_handling, :event_groups)
-      next unless provider_event_groups
-      DeepMerge.deep_merge!(
-        provider_event_groups.to_hash, event_groups,
-        :preserve_unmergeables => false,
-        :overwrite_arrays      => false
-      )
+    @event_groups ||= begin # TODO: this cache needs to be busted whenever settings are changed (and also in test)
+      core_event_groups = ::Settings.event_handling.event_groups.to_hash
+      Settings.ems.each_with_object(core_event_groups) do |(_provider_type, provider_settings), event_groups|
+        provider_event_groups = provider_settings.fetch_path(:event_handling, :event_groups)
+        next unless provider_event_groups
+
+        DeepMerge.deep_merge!(
+          provider_event_groups.to_hash, event_groups,
+          :preserve_unmergeables => false,
+          :overwrite_arrays      => false
+        )
+      end
     end
+  end
+
+  private_class_method def self.partition_group_and_level_by_event_type
+    # TODO: this cache needs to be busted whenever settings are changed (and also in test)
+    return @literal_group_and_level_by_event_type, @regex_group_and_level_by_event_type if @literal_group_and_level_by_event_type
+
+    @literal_group_and_level_by_event_type = {}
+    @regex_group_and_level_by_event_type = {}
+
+    event_groups.each do |group_name, group_contents|
+      group_contents.each do |group_level, event_types|
+        next if group_level == :name
+
+        event_types.each do |event_type|
+          if event_type.starts_with?("/")
+            @regex_group_and_level_by_event_type[Regexp.new(event_type[1..-2])] ||= [group_name, group_level]
+          else
+            @literal_group_and_level_by_event_type[event_type] ||= [group_name, group_level]
+          end
+        end
+      end
+    end
+
+    return @literal_group_and_level_by_event_type, @regex_group_and_level_by_event_type
+  end
+
+  def self.clear_event_groups_cache
+    @event_groups = @literal_group_and_level_by_event_type = @regex_group_and_level_by_event_type = nil
   end
 
   # TODO: Consider moving since this is EmsEvent specific. group, group_level and group_name exposed as a virtual columns for reports/api.
   def self.group_and_level(event_type)
-    level = DEFAULT_GROUP_LEVEL # the level is detail as default
-    egroups = event_groups
-
-    group = egroups.detect do |_, value|
-      group_levels
-        .detect { |lvl| value[lvl]&.any? { |typ| !typ.starts_with?("/") && typ == event_type } }
-        .tap { |level_found| level = level_found || level }
-    end&.first
-
-    group ||= egroups.detect do |_, value|
-      group_levels
-        .detect { |lvl| value[lvl]&.any? { |typ| typ.starts_with?("/") && Regexp.new(typ[1..-2]).match?(event_type) } }
-        .tap { |level_found| level = level_found || level }
-    end&.first
-
-    group ||= DEFAULT_GROUP_NAME
-    return group, level
+    by_literal, by_regex = partition_group_and_level_by_event_type
+    by_literal[event_type] ||
+      by_regex.detect { |regex, _| regex.match?(event_type) }&.last ||
+      [DEFAULT_GROUP_NAME, DEFAULT_GROUP_LEVEL]
   end
 
   def self.group_name(group)
     return if group.nil?
-    group = event_groups[group.to_sym]
-    group.nil? ? DEFAULT_GROUP_NAME.to_s.capitalize : group[:name]
+
+    event_groups.dig(group.to_sym, :name) || DEFAULT_GROUP_NAME_STR
   end
 
   # TODO: Consider moving since this is EmsEvent specific. group, group_level and group_name exposed as a virtual columns for reports/api.
   def group
-    return @group unless @group.nil?
-    @group, @group_level = self.class.group_and_level(event_type)
-    @group
+    group_and_level.first
   end
 
   # TODO: Consider moving since this is EmsEvent specific. group, group_level and group_name exposed as a virtual columns for reports/api.
   def group_level
-    return @group_level unless @group_level.nil?
-    @group, @group_level = self.class.group_and_level(event_type)
-    @group_level
+    group_and_level.last
+  end
+
+  private def group_and_level
+    @group_and_level ||= self.class.group_and_level(event_type)
   end
 
   # TODO: Consider moving since this is EmsEvent specific. group, group_level and group_name exposed as a virtual columns for reports/api.