Based on #22304 

The way this works is we take the resolved set of event_groups and cache
that. This should not change unless the settings change.

After that, instead of always walking the event_groups during lookup, we
walk them once and partition it into one lookup table for String literals
and another lookup table for Regexes, and cache those. Later, when we
want to find the group_and_level we go to those lookup tables instead.

This drops the old lookup code from O(N) + a lot of allocated objects to
O(1 String) + O(N Regex) (note that there are many Strings and very few
Regex).

@jrafanie Please review
cc @kbrock @MelsHyrule 

~~WIP: I still need to bust the caches when Settings changes~~ DONE

---

Before: The following example code took forever (I canceled it and never let it complete)
After: The group, group_name, and group_level lookups are sub-millisecond

```
# With 300k event records

# Baseline query to determine query, instantiation, and Array conversion overhead
irb(main):001:0> Benchmark.ms { EmsEvent.all.to_a; nil }
  EmsEvent Load (711.7ms)  SELECT "event_streams".* FROM "event_streams" WHERE "event_streams"."type" = $1  [["type", "EmsEvent"]]
  EmsEvent Inst Including Associations (6830.4ms - 297116rows)
=> 14751.755999983288

irb(main):002:0> Benchmark.ms { EmsEvent.all.each { |e| [e.group, e.group_name, e.group_level] }; nil }
  EmsEvent Load (756.9ms)  SELECT "event_streams".* FROM "event_streams" WHERE "event_streams"."type" = $1  [["type", "EmsEvent"]]
  EmsEvent Inst Including Associations (6682.2ms - 297116rows)
=> 15410.467999987304
```

---

Future PR - After this we have basically lookup tables by `event_type -> [group, group_level]`, so if we want to query by group or group_level we can just invert that.  I believe it shouldn't be too difficult to make it a scope, and that would allow filtering in SQL.