@jrafanie 2023-01-31
It's hard for me to follow the changes here all in one commit.  Is this still WIP?  I'd like to close https://github.com/ManageIQ/manageiq/pull/22304 if this is the better replacement.