@Fryguy 2023-02-06
It's only one change, hence only one commit - I didn't include any of the smaller optimizations from #22304.

I can do a live review if you'd like.  This is WIP because we'd need to include cache busting on setting change, which I haven't had a chance to implement yet.