@Fryguy 2023-02-07
Removed the commit for the scopes as that was a PoC.